pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/debussy"]

[project]
name = "claude-debussy"
version = "0.1.0"
description = "Orchestrate multi-phase Claude CLI sessions with compliance verification"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "typer>=0.21.1",
]

[project.scripts]
debussy = "debussy.cli:app"

[dependency-groups]
dev = [
    "bandit>=1.9.2",
    "ltm",
    "pre-commit>=4.5.1",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "radon>=6.0.1",
    "ruff>=0.14.11",
    "ty>=0.0.11",
]

[tool.uv.sources]
ltm = { url = "https://github.com/matt-grain/Claude-LTM/releases/download/v0.4.5/ltm-0.4.5-py3-none-any.whl" }

# ============================================================================
# Tool Configuration - Eating Our Own Dog Food
# ============================================================================

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented code)
    "PL",     # Pylint
    "RUF",    # Ruff-specific
]
ignore = [
    "PLR0913",  # Too many arguments (we have config objects)
    "PLR2004",  # Magic value comparison (fine in tests)
    "TC003",    # Path used at runtime in Pydantic models
    "PLC0415",  # Intentional lazy imports to avoid circular deps
]

[tool.ruff.lint.isort]
known-first-party = ["debussy"]

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "standard"  # strict is too aggressive with Pydantic
include = ["src"]
exclude = ["**/__pycache__", ".venv"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "-v",
    "--cov=debussy",
    "--cov-report=term-missing",
    "--cov-fail-under=50",  # ClaudeRunner streaming and CLI banner hard to unit test
]

[tool.coverage.run]
source = ["src/debussy"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101"]  # assert is fine in non-production code
